Basic gameplay, bug fixes, design
Created constructor for the Enemy class;
colored the map and the current view of the player to make it more appealing;
fixed out ofrange bugs (due to jagged arrays);
monsters and items now disappear when dealt with.

Fix: The player doesn't receive XP after killing a monster.
Add: Make title and parts of texts more colorful.